third_party/WebKit/LayoutTests/css3/device-adapt/opera/syntax-001-expected.txt


FAIL CSS Test: @viewport syntax - inside @media assert_equals: expected 500 but got 800
